dagster-skills-evals/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dagster_skills_evals"
version = "0.0.6"
description = "Utilities for evaluating the dagster-provided skills"
requires-python = ">=3.13,<3.14"
dependencies = [
    "dagster_shared",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "typer>=0.15.0",
]

[project.scripts]
dagster-skills = "dagster_skills_evals.cli:app"

[project.optional-dependencies]
test = [
    "dagster",
    "dagster-duckdb",
    "dagster-fivetran",
    "dagster-dbt",
    "markdown-it-py",
    "pandas",
    "pyright",
    "pytest>=8.0.0",
    "ruff",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "F",      # Pyflakes
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "YTT",    # flake8-2020
    "ASYNC",  # flake8-async
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # Pylint
    "TRY",    # tryceratops
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",    # Line too long (handled by formatter)
    "ISC001",  # Implicit string concatenation (conflicts with formatter)
    "PLR0913", # Too many arguments to function call
    "TRY003",  # Avoid specifying long messages outside the exception class
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py files
"**/cli/*.py" = ["B008"]  # typer.Argument/Option in defaults is the intended pattern

[tool.ruff.lint.isort]
known-first-party = ["dagster_skills_evals"]
